backend/requirements.txt
flask
flask-socketio
eventlet
python-socketio
gunicorn
python-dotenv
